Factor subclasses of RDD out of RDD.scala into their own classes in the rdd package.
