+++ core/src/main/scala/spark/rdd/MappedRDD.scala
@@ -1,0 +1,16 @@
+package spark.rdd
+
+import spark.OneToOneDependency
+import spark.RDD
+import spark.Split
+
+private[spark]
+class MappedRDD[U: ClassManifest, T: ClassManifest](
+    prev: RDD[T],
+    f: T => U)
+  extends RDD[U](prev.context) {
+  
+  override def splits = prev.splits
+  override val dependencies = List(new OneToOneDependency(prev))
+  override def compute(split: Split) = prev.iterator(split).map(f)
+}
